Aggiustati i dizionari. Creato il plot che tanto volevo. Aggiunta utility per salvare i plots.
